+++ backend/Dockerfile
@@ -14,5 +14,7 @@ RUN pip install --no-cache-dir -r requirements.txt
 # Copy application code
 COPY main.py .
 COPY analytics ./analytics
+COPY integrations ./integrations
+COPY bitcoin_utils.py .
 
 CMD ["python", "main.py"]
