+++ backend/main.py
@@ -1,16 +1,8 @@
 import os, time, threading, logging
-from fastapi import FastAPI
-from bitcoinrpc.authproxy import AuthServiceProxy, JSONRPCException
-from pymongo import MongoClient
-from pydantic import BaseModel
-import uvicorn
+from bitcoinrpc.authproxy import AuthServiceProxy
 from io import BytesIO
 import zmq
-import json
-import pika
-import requests
 import uuid
-import re
 from datetime import datetime
 import concurrent.futures
 from queue import Queue, Empty
@@ -19,12 +11,13 @@ from threading import Lock
 import random
 from distutils.util import strtobool
 from typing import List, Dict, Any
+from bitcoin_utils import extract_coinbase_data
+from integrations import db, blocks_coll, pools_coll, mongodb_enabled, publish_to_rabbitmq, rabbitmq_manager
+from integrations.rabbitmq import start_heartbeat_thread
 
-# For decoding transactions when running analysis
-try:
-    from bitcoin.core import CTransaction
-except Exception:
-    CTransaction = None  # Analysis that requires tx parsing will be skipped if unavailable
+# Global mining pool definitions cache
+pools_cache: Dict[str, Dict[str, Any]] = {}
+pools_hash: int | None = None
 
 # --- Logging Setup ---
 logging.basicConfig(
@@ -34,50 +27,33 @@ logging.basicConfig(
 logger = logging.getLogger("backend")
 
 # --- Configuration (parameterized via environment variables) ---
-ENABLE_HISTORICAL_DATA_STR = os.getenv('ENABLE_HISTORICAL_DATA', 'true')
-try:
-    ENABLE_HISTORICAL_DATA = bool(strtobool(ENABLE_HISTORICAL_DATA_STR))
-except ValueError:
-    logger.error(f"Invalid value for ENABLE_HISTORICAL_DATA: {ENABLE_HISTORICAL_DATA_STR}. Defaulting to True.")
-    ENABLE_HISTORICAL_DATA = True
-
-logger.info(f"Historical data feature enabled: {ENABLE_HISTORICAL_DATA}")
 
 RPC_USER = os.getenv("BITCOIN_RPC_USER", "user")
 RPC_PASSWORD = os.getenv("BITCOIN_RPC_PASSWORD", "password")
 RPC_HOST = os.getenv("BITCOIN_RPC_HOST", "localhost")
 RPC_PORT = os.getenv("BITCOIN_RPC_PORT", "8332")
 RPC_URL = f"http://{RPC_USER}:{RPC_PASSWORD}@{RPC_HOST}:{RPC_PORT}"
+# Timeout in seconds for individual RPC HTTP operations
+RPC_TIMEOUT = float(os.getenv("BITCOIN_RPC_TIMEOUT", "10"))
 logger.info("Connecting to bitcoin RPC at %s:%s", RPC_HOST, RPC_PORT)
 
 ZMQ_BLOCK = os.getenv("BITCOIN_ZMQ_BLOCK", "tcp://bitcoin-node:28332")
 MIN_BLOCK_HEIGHT = int(os.getenv("MIN_BLOCK_HEIGHT", "882000"))
 
-MONGO_URL = os.getenv("MONGODB_URL", "mongodb://mongodb:27017")
-MONGO_DB = os.getenv("MONGODB_DB", "stratum-logger")
-MONGO_USER = os.getenv("MONGODB_USERNAME", "mongouser")
-MONGO_PASSWORD = os.getenv("MONGODB_PASSWORD", "mongopassword")
-
-RABBITMQ_HOST = os.getenv("RABBITMQ_HOST", "rabbitmq")
-RABBITMQ_PORT = int(os.getenv("RABBITMQ_PORT", "5672"))
-RABBITMQ_USERNAME = os.getenv("RABBITMQ_USERNAME", "mquser")
-RABBITMQ_PASSWORD = os.getenv("RABBITMQ_PASSWORD", "mqpassword")
-RABBITMQ_EXCHANGE = os.getenv("RABBITMQ_EXCHANGE", "blocks")
-RABBITMQ_HEARTBEAT = int(os.getenv("RABBITMQ_HEARTBEAT", "30"))  # Reduced from 60 to 30
-RABBITMQ_CONNECTION_TIMEOUT = int(os.getenv("RABBITMQ_CONNECTION_TIMEOUT", "10"))
-RABBITMQ_SOCKET_TIMEOUT = int(os.getenv("RABBITMQ_SOCKET_TIMEOUT", "5"))
-RABBITMQ_RETRY_DELAY = int(os.getenv("RABBITMQ_RETRY_DELAY", "2"))
-RABBITMQ_MAX_RETRIES = int(os.getenv("RABBITMQ_MAX_RETRIES", "5"))
+# MongoDB configuration is managed inside integrations.mongodb
 
 POOL_LIST_URL = os.getenv("POOL_LIST_URL", "https://raw.githubusercontent.com/mempool/mining-pools/refs/heads/master/pools-v2.json")
 POOL_UPDATE_INTERVAL = int(os.getenv("POOL_UPDATE_INTERVAL", "86400"))  # Default: once per day
 LOCAL_POOL_FILE = os.getenv("LOCAL_POOL_FILE", "pool_definitions.json")  # Local fallback file
 
+# Controls whether to start a full background reindex on startup
+START_FULL_REINDEX_ON_BOOT = False
+
 # --- Connections ---
 def create_rpc_connection():
     for attempt in range(1, 6):
         try:
-            conn = AuthServiceProxy(RPC_URL)
+            conn = AuthServiceProxy(RPC_URL, timeout=RPC_TIMEOUT)
             conn.getblockcount()
             logger.info("Successfully connected to bitcoin RPC (attempt %d)", attempt)
             return conn
@@ -85,7 +61,7 @@ def create_rpc_connection():
             logger.error("RPC connection attempt %d failed: %s", attempt, e)
             # Reinitialize connection (do not log credentials)
             try:
-                conn = AuthServiceProxy(RPC_URL)
+                conn = AuthServiceProxy(RPC_URL, timeout=RPC_TIMEOUT)
                 logger.info("Reinitialized bitcoin RPC connection on attempt %d", attempt)
                 return conn
             except Exception as reinit_err:
@@ -95,291 +71,15 @@ def create_rpc_connection():
 
 bitcoin_rpc = create_rpc_connection()
 
-mongo_client = None
-db = None
-blocks_coll = None
-pools_coll = None
-
-if ENABLE_HISTORICAL_DATA:
-    logger.info("Historical data enabled, attempting to connect to MongoDB...")
-    try:
-        mongo_client = MongoClient(MONGO_URL, username=MONGO_USER, password=MONGO_PASSWORD)
-        # Verify connection
-        mongo_client.admin.command('ping') 
-        db = mongo_client[MONGO_DB]
-        blocks_coll = db.blocks
-        pools_coll = db.pools
-        logger.info("Successfully connected to MongoDB at %s", MONGO_URL)
-    except Exception as e:
-        logger.error("Error connecting to MongoDB: %s. Historical data features will be disabled.", e)
-        # Ensure all are None if connection fails
-        mongo_client = None 
-        db = None
-        blocks_coll = None
-        pools_coll = None
-        ENABLE_HISTORICAL_DATA = False # Explicitly disable if connection fails
-else:
-    logger.info("Historical data disabled, skipping MongoDB connection.")
-
-# --- RabbitMQ Connection Management ---
-class RabbitMQManager:
-    def __init__(self, host, port, username, password, exchange, 
-                 heartbeat=RABBITMQ_HEARTBEAT, 
-                 connection_timeout=RABBITMQ_CONNECTION_TIMEOUT,
-                 socket_timeout=RABBITMQ_SOCKET_TIMEOUT):
-        self.host = host
-        self.port = port
-        self.username = username
-        self.password = password
-        self.exchange = exchange
-        self.heartbeat = heartbeat
-        self.connection_timeout = connection_timeout
-        self.socket_timeout = socket_timeout
-        self.connection = None
-        self.channel = None
-        self.lock = Lock()
-        self.last_reconnect_attempt = 0
-        self.reconnect_cooldown = 2
-        self.connect()
-    
-    def connect(self):
-        """Establish a connection to RabbitMQ"""
-        with self.lock:
-            # Close existing connection if it exists
-            if self.connection:
-                try:
-                    if not self.connection.is_closed:
-                        self.connection.close()
-                except Exception as e:
-                    logger.warning(f"Error closing existing RabbitMQ connection: {e}")
-            
-            max_retries = RABBITMQ_MAX_RETRIES
-            retry_delay = RABBITMQ_RETRY_DELAY
-            
-            for attempt in range(max_retries):
-                try:
-                    # Use URLParameters for more robust connection handling
-                    url = f"amqp://{self.username}:{self.password}@{self.host}:{self.port}/%2F"
-                    params = pika.URLParameters(url)
-                    
-                    # Override default parameters with our custom settings
-                    params.heartbeat = self.heartbeat
-                    params.socket_timeout = self.socket_timeout
-                    params.blocked_connection_timeout = self.connection_timeout
-                    params.retry_delay = 1
-                    params.connection_attempts = 3
-                    
-                    # Add TCP keepalive settings to prevent connection drops
-                    params.tcp_options = {
-                        'TCP_KEEPIDLE': 60,  # Start sending keepalive probes after 60 seconds
-                        'TCP_KEEPINTVL': 10,  # Send keepalive probes every 10 seconds
-                        'TCP_KEEPCNT': 5      # Drop connection after 5 failed probes
-                    }
-                    
-                    logger.info(f"Connecting to RabbitMQ at {self.host}:{self.port} (heartbeat={self.heartbeat}s)")
-                    self.connection = pika.BlockingConnection(params)
-                    
-                    self.channel = self.connection.channel()
-                    # Set prefetch count to 1 to avoid overwhelming the channel
-                    self.channel.basic_qos(prefetch_count=1)
-                    self.channel.exchange_declare(
-                        exchange=self.exchange, 
-                        exchange_type='fanout', 
-                        durable=True
-                    )
-                    
-                    logger.info(f"Successfully connected to RabbitMQ at {self.host}:{self.port}")
-                    self.last_reconnect_attempt = time.time()
-                    return True
-                except (pika.exceptions.AMQPConnectionError, 
-                        pika.exceptions.ConnectionClosedByBroker,
-                        pika.exceptions.ConnectionWrongStateError,
-                        pika.exceptions.StreamLostError) as e:
-                    logger.error(f"Failed to connect to RabbitMQ (attempt {attempt+1}/{max_retries}): {e}")
-                    if attempt < max_retries - 1:
-                        time.sleep(retry_delay)
-                        retry_delay = min(retry_delay * 1.5, 10)  # Exponential backoff with cap
-                except Exception as e:
-                    logger.error(f"Unexpected error connecting to RabbitMQ (attempt {attempt+1}/{max_retries}): {e}")
-                    if attempt < max_retries - 1:
-                        time.sleep(retry_delay)
-                        retry_delay = min(retry_delay * 1.5, 10)  # Exponential backoff with cap
-            
-            logger.error("Failed to connect to RabbitMQ after all retries")
-            return False
-    
-    def ensure_connection(self):
-        """Ensure that the connection is active, reconnect if needed"""
-        current_time = time.time()
-        
-        # Check if connection is closed or doesn't exist
-        if not self.connection or self.connection.is_closed:
-            # Check if we're in cooldown period
-            if current_time - self.last_reconnect_attempt < self.reconnect_cooldown:
-                logger.warning("RabbitMQ reconnection in cooldown period, skipping reconnect attempt")
-                return False
-            
-            logger.info("RabbitMQ connection is closed, attempting to reconnect")
-            return self.connect()
-        
-        # Check if connection is still alive
-        try:
-            # Process any pending events to keep the connection alive
-            self.connection.process_data_events(time_limit=0)
-        except (pika.exceptions.ConnectionClosed, 
-                pika.exceptions.ChannelClosed, 
-                pika.exceptions.AMQPConnectionError,
-                pika.exceptions.StreamLostError) as e:
-            logger.warning(f"RabbitMQ connection error during health check: {e}")
-            return self.connect()
-        except Exception as e:
-            logger.warning(f"Error checking RabbitMQ connection health: {e}")
-            return self.connect()
-            
-        return True
-    
-    def publish(self, doc, max_retries=3):
-        """Publish a message to RabbitMQ with retry logic"""
-        retry_delay = RABBITMQ_RETRY_DELAY
-        message_type = doc.get('type', 'unknown')
-        
-        for attempt in range(max_retries):
-            try:
-                # Ensure connection is active
-                if not self.ensure_connection():
-                    if attempt < max_retries - 1:
-                        logger.warning(f"Failed to ensure RabbitMQ connection, retrying in {retry_delay}s")
-                        time.sleep(retry_delay)
-                        retry_delay = min(retry_delay * 1.5, 10)  # Exponential backoff with cap
-                        continue
-                    else:
-                        logger.error("Failed to publish: could not establish RabbitMQ connection")
-                        return False
-                
-                # Prepare message
-                message_body = json.dumps(doc)
-                
-                # Publish message with more robust error handling
-                try:
-                    self.channel.basic_publish(
-                        exchange=self.exchange,
-                        routing_key='',
-                        body=message_body,
-                        properties=pika.BasicProperties(
-                            delivery_mode=2,  # make message persistent
-                            content_type='application/json',
-                            message_id=str(uuid.uuid4()),  # Add unique message ID
-                            timestamp=int(time.time())     # Add timestamp
-                        ),
-                        mandatory=False  # Don't require confirmation
-                    )
-                    
-                    # Process any pending events to keep the connection alive
-                    self.connection.process_data_events(time_limit=0)
-                    
-                    logger.info(f"Published message to RabbitMQ: {message_type}")
-                    return True
-                except (pika.exceptions.UnroutableError, pika.exceptions.NackError) as e:
-                    # These are specific publishing errors
-                    logger.error(f"Message routing error (attempt {attempt+1}/{max_retries}): {e}")
-                    # Force reconnection on next attempt
-                    self.connection = None
-                    self.channel = None
-            
-            except (pika.exceptions.ConnectionClosed, 
-                    pika.exceptions.ChannelClosed, 
-                    pika.exceptions.AMQPConnectionError,
-                    pika.exceptions.StreamLostError) as e:
-                logger.error(f"RabbitMQ connection error (attempt {attempt+1}/{max_retries}): {e}")
-                # Force reconnection on next attempt
-                self.connection = None
-                self.channel = None
-                
-                if attempt < max_retries - 1:
-                    time.sleep(retry_delay)
-                    retry_delay = min(retry_delay * 1.5, 10)  # Exponential backoff with cap
-            
-            except Exception as e:
-                logger.error(f"Unexpected error publishing to RabbitMQ (attempt {attempt+1}/{max_retries}): {e}")
-                # For unexpected errors, try to reconnect as well
-                self.connection = None
-                self.channel = None
-                
-                if attempt < max_retries - 1:
-                    time.sleep(retry_delay)
-                    retry_delay = min(retry_delay * 1.5, 10)  # Exponential backoff with cap
-                else:
-                    logger.error("Failed to publish to RabbitMQ after all retries")
-                    return False
-        
-        return False
-
-# Initialize RabbitMQ manager
-rabbitmq_manager = RabbitMQManager(
-    host=RABBITMQ_HOST,
-    port=RABBITMQ_PORT,
-    username=RABBITMQ_USERNAME,
-    password=RABBITMQ_PASSWORD,
-    exchange=RABBITMQ_EXCHANGE
-)
-
-# For backward compatibility, provide a publish_to_rabbitmq function
-def publish_to_rabbitmq(doc):
-    return rabbitmq_manager.publish(doc)
-
-# --- FastAPI app and models ---
-app = FastAPI()
-
-# Create separate thread pools for old and new blocks
-old_block_processor = concurrent.futures.ThreadPoolExecutor(max_workers=4, thread_name_prefix="old_block_processor")
-new_block_processor = concurrent.futures.ThreadPoolExecutor(max_workers=2, thread_name_prefix="new_block_processor")
-
-class BlockModel(BaseModel):
-    height: int
-    coinbase_script_sig: str
-    block_hash: str
-    timestamp: int
-
-# Event to signal that block sync is complete.
-sync_complete_event = threading.Event()
-
-@app.get("/blocks")
-def get_blocks(n: int = 20, before: int = None):
-    try:
-        # Build the query based on the before parameter
-        query = {}
-        if before is not None:
-            # Include the 'before' height in the results to avoid gaps
-            query["height"] = {"$lte": before}
-        
-        # Get one more block than requested to determine if there are more blocks
-        blocks = list(blocks_coll.find(query, {"_id": 0}).sort("height", -1).limit(n + 1))
-        
-        # Check if we have more blocks
-        has_more = len(blocks) > n
-        if has_more:
-            # Remove the extra block we fetched
-            blocks = blocks[:n]
-        
-        # Log the block heights for debugging
-        heights = [block["height"] for block in blocks]
-        logger.info("API call /blocks succeeded: %d blocks returned, heights: %s", len(blocks), heights)
-        
-        # Return both the blocks and pagination info
-        return {
-            "blocks": blocks,
-            "has_more": has_more,
-            "next_height": blocks[-1]["height"] if blocks else None
-        }
-    except Exception as e:
-        logger.error("API call /blocks failed: %s", e)
-        return {"error": str(e)}
+# Create thread pool for block processing (ZMQ real-time handling)
+block_processor = concurrent.futures.ThreadPoolExecutor(max_workers=2, thread_name_prefix="block_processor")
 
 # --- RPC Connection Pool ---
 class RPCConnectionPool:
-    def __init__(self, url, pool_size=4):
+    def __init__(self, url, pool_size=4, rpc_timeout: float = RPC_TIMEOUT):
         self.url = url
         self.pool_size = pool_size
+        self.rpc_timeout = rpc_timeout
         self.connection_queue = Queue()
         self.lock = Lock()
         self.last_recreation_time = 0
@@ -387,106 +87,41 @@ class RPCConnectionPool:
         self._initialize_pool()
     
     def _initialize_pool(self):
-        """Initialize the connection pool with fresh connections"""
+        """Initialize the pool; connections are created per RPC call."""
         with self.lock:
-            # Clear existing queue
             while not self.connection_queue.empty():
                 try:
                     self.connection_queue.get_nowait()
                 except Empty:
                     break
-            
-            # Create new connections
-            for _ in range(self.pool_size):
-                try:
-                    conn = AuthServiceProxy(self.url)
-                    # Test the connection with a simple call
-                    conn.getblockcount()
-                    self.connection_queue.put(conn)
-                except Exception as e:
-                    logger.error(f"Failed to create RPC connection: {e}")
-            
-            if self.connection_queue.empty():
-                logger.error("Failed to create any RPC connections")
+            logger.info("RPC pool initialized; connections are created per RPC call")
     
-    def recreate_pool(self):
+    def recreate_pool(self, force: bool = False):
         """Recreate the entire connection pool"""
         current_time = time.time()
         
         # Check if we're in cooldown period
-        if current_time - self.last_recreation_time < self.recreation_cooldown:
+        if not force and (current_time - self.last_recreation_time < self.recreation_cooldown):
             logger.warning("RPC pool recreation in cooldown period, skipping recreation")
             return False
         
         with self.lock:
-            logger.warning("Recreating entire RPC connection pool")
-            
-            # Close existing connections
-            while not self.connection_queue.empty():
-                try:
-                    conn = self.connection_queue.get_nowait()
-                    # No explicit close method for AuthServiceProxy, but we can 
-                    # remove references to allow garbage collection
-                    del conn
-                except Empty:
-                    break
-            
-            # Create new connections
-            success = False
-            for _ in range(self.pool_size):
-                try:
-                    conn = AuthServiceProxy(self.url)
-                    # Test the connection with a simple call
-                    conn.getblockcount()
-                    self.connection_queue.put(conn)
-                    success = True
-                except Exception as e:
-                    logger.error(f"Failed to create RPC connection during recreation: {e}")
-            
+            logger.warning("Recreating RPC connection pool")
             self.last_recreation_time = current_time
-            
-            if success:
-                logger.info("Successfully recreated RPC connection pool")
-                return True
-            else:
-                logger.error("Failed to recreate any RPC connections")
-                return False
+            logger.info("RPC pool recreation complete")
+            return True
     
     @contextmanager
     def get_connection(self):
         """Get a connection from the pool"""
         conn = None
         try:
-            # Try to get a connection with timeout
-            try:
-                conn = self.connection_queue.get(timeout=2)
-            except Empty:
-                # If queue is empty, try to recreate the pool
-                logger.warning("RPC connection queue is empty, attempting to recreate")
-                self.recreate_pool()
-                try:
-                    conn = self.connection_queue.get(timeout=2)
-                except Empty:
-                    logger.error("Failed to get RPC connection after recreation")
-                    raise Exception("Error getting RPC connection")
-            
-            # Test connection before returning it
-            try:
-                # Simple ping test
-                conn.getblockcount()
-                yield conn
-            except Exception as e:
-                logger.error(f"RPC connection test failed: {e}")
-                # Don't return this connection to the pool
-                self.recreate_pool()
-                raise
-        except Exception as e:
-            logger.error(f"Error getting RPC connection: {e}")
-            raise
-        else:
-            # Return connection to the pool if it's still valid
-            if conn:
-                self.connection_queue.put(conn)
+            # Create a fresh connection per use to avoid stale keep-alive sockets
+            conn = AuthServiceProxy(self.url, timeout=self.rpc_timeout)
+            yield conn
+        finally:
+            # No persistent connection to return; allow garbage collection
+            conn = None
 
 # Create the connection pool
 rpc_pool = RPCConnectionPool(RPC_URL)
@@ -542,7 +177,7 @@ def retry_rpc(func, description, *args, attempts=5, initial_delay=2, **kwargs):
                     ["connection", "timeout", "refused", "reset", "broken pipe", "eof"]):
                 logger.warning("Detected RPC connection issue, attempting to recreate connection pool")
                 # Force recreation of the connection pool
-                rpc_pool.recreate_pool()
+                rpc_pool.recreate_pool(force=True)
                 # Add a delay to allow connections to stabilize
                 time.sleep(delay)
             
@@ -564,122 +199,71 @@ def retry_rpc(func, description, *args, attempts=5, initial_delay=2, **kwargs):
 
 
 # --- Block processing ---
+def build_block_doc(block_hash: str) -> Dict[str, Any]:
+    block = retry_rpc(
+        get_block,
+        f"getblock({block_hash})",
+        block_hash
+    )
+    height = block["height"]
+    coinbase_script_sig, coinbase_addresses = extract_coinbase_data(block)
+
+    analysis = {}
+    try:
+        if mongodb_enabled and db is not None:
+            analysis = run_block_analyses(height, coinbase_script_sig, coinbase_addresses)
+    except Exception as analysis_err:
+        logger.error(f"Error running analyses for height {height}: {analysis_err}")
+        analysis = {}
+
+    mining_pool = {}
+    try:
+        if isinstance(analysis, dict) and analysis.get("pool_identification"):
+            pool_info = analysis.get("pool_identification", {})
+            mining_pool = pool_info.get("mining_pool", {}) or {}
+            logger.info(
+                "Pool identification via analysis: name=%s id=%s method=%s",
+                mining_pool.get('name', 'Unknown'),
+                mining_pool.get('id', 'unknown'),
+                pool_info.get('method', 'unknown')
+            )
+        else:
+            mining_pool = identify_pool_from_data(pools_cache, coinbase_script_sig, coinbase_addresses)
+            logger.info(
+                "Pool identification via fallback: name=%s id=%s",
+                mining_pool.get('name', 'Unknown'),
+                mining_pool.get('id', 'unknown')
+            )
+    except Exception as pool_err:
+        logger.error(f"Error identifying mining pool at height {height}: {pool_err}")
+        mining_pool = {"name": "Unknown", "id": "unknown"}
+
+    doc = {
+        "height": height,
+        "coinbase_script_sig": coinbase_script_sig,
+        "block_hash": block_hash,
+        "timestamp": block["time"],
+        "mining_pool": mining_pool,
+        "analysis": analysis
+    }
+    return doc
+
 def process_block(block_hash: str, is_new_block: bool = False):
     try:
-        # Use a shorter timeout for new blocks to ensure responsiveness
-        timeout = 30 if is_new_block else 120
-        
-        block = retry_rpc(
-            get_block, 
-            f"getblock({block_hash})",
-            block_hash
-        )
-        height = block["height"]
-        coinbase_tx = block["tx"][0]
-        vin0 = coinbase_tx["vin"][0]
-        if "scriptSig" in vin0 and "hex" in vin0["scriptSig"]:
-            coinbase_script_sig = vin0["scriptSig"]["hex"]
-        elif "coinbase" in vin0:
-            coinbase_script_sig = vin0["coinbase"]
-        else:
-            raise KeyError("No coinbase script found")
-            
-        # Extract coinbase addresses - improved to match the Rust implementation
-        coinbase_addresses = []
-        for out in coinbase_tx["vout"]:
-            script_pub_key = out.get("scriptPubKey", {})
-            
-            # First check for addresses array
-            if "addresses" in script_pub_key:
-                coinbase_addresses.extend(script_pub_key["addresses"])
-            # Then check for single address
-            elif "address" in script_pub_key:
-                coinbase_addresses.append(script_pub_key["address"])
-            # For newer Bitcoin Core versions that use 'address' field with descriptor
-            elif "desc" in script_pub_key and "address" in script_pub_key:
-                coinbase_addresses.append(script_pub_key["address"])
-        
-        # Sort addresses by value (descending) to match Rust implementation
-        # This requires getting the values from the vout
-        address_values = {}
-        for out in coinbase_tx["vout"]:
-            script_pub_key = out.get("scriptPubKey", {})
-            value = out.get("value", 0)
-            
-            addresses = []
-            if "addresses" in script_pub_key:
-                addresses = script_pub_key["addresses"]
-            elif "address" in script_pub_key:
-                addresses = [script_pub_key["address"]]
-            
-            for addr in addresses:
-                if addr in address_values:
-                    address_values[addr] += value
-                else:
-                    address_values[addr] = value
-        
-        # Sort addresses by value (descending)
-        coinbase_addresses = sorted(
-            coinbase_addresses,
-            key=lambda addr: address_values.get(addr, 0),
-            reverse=True
-        )
-        
-        # Run composable analyses using mining.notify templates for this height
-        analysis = {}
-        try:
-            if ENABLE_HISTORICAL_DATA and db is not None:
-                analysis = run_block_analyses(height, coinbase_script_sig, coinbase_addresses)
-        except Exception as analysis_err:
-            logger.error(f"Error running analyses for height {height}: {analysis_err}")
-            analysis = {}
-
-        # Determine mining pool from analysis (preferred), otherwise fallback to Unknown
-        mining_pool = {}
-        try:
-            if isinstance(analysis, dict) and analysis.get("pool_identification"):
-                pool_info = analysis.get("pool_identification", {})
-                mining_pool = pool_info.get("mining_pool", {}) or {}
-                logger.info(
-                    "Pool identification via analysis: name=%s id=%s method=%s",
-                    mining_pool.get('name', 'Unknown'),
-                    mining_pool.get('id', 'unknown'),
-                    pool_info.get('method', 'unknown')
-                )
-            else:
-                # Fallback to direct identification if analysis did not produce it
-                mining_pool = pools_manager.identify_pool(coinbase_script_sig, coinbase_addresses)
-                logger.info(
-                    "Pool identification via fallback: name=%s id=%s",
-                    mining_pool.get('name', 'Unknown'),
-                    mining_pool.get('id', 'unknown')
-                )
-        except Exception as pool_err:
-            logger.error(f"Error identifying mining pool at height {height}: {pool_err}")
-            mining_pool = {"name": "Unknown", "id": "unknown"}
-
-        doc = {
-            "height": height,
-            "coinbase_script_sig": coinbase_script_sig,
-            "block_hash": block_hash,
-            "timestamp": block["time"],
-            "mining_pool": mining_pool,
-            "analysis": analysis
-        }
-        
+        doc = build_block_doc(block_hash)
+        height = doc["height"]
         # For new blocks, use update_one with upsert to avoid duplicates
         # For old blocks during sync, use insert_one since we already checked for existence
         if is_new_block:
             blocks_coll.update_one({"block_hash": block_hash}, {"$set": doc}, upsert=True)
         else:
             blocks_coll.insert_one(doc)
-            
-        logger.info("%s block %d (%s) mined by %s", 
-                   "Processed new" if is_new_block else "Synced old", 
-                   height, block_hash, 
-                   mining_pool.get('name') if mining_pool else 'Unknown')
 
-        # Publish block update to RabbitMQ with type information
+        logger.info("%s block %d (%s) mined by %s",
+                   "Processed new" if is_new_block else "Synced old",
+                   height, block_hash,
+                   (doc.get("mining_pool") or {}).get('name') if doc.get("mining_pool") else 'Unknown')
+
         rabbitmq_doc = {
             "type": "block",
             "id": str(uuid.uuid4()),
@@ -687,21 +271,61 @@ def process_block(block_hash: str, is_new_block: bool = False):
             "data": {
                 "hash": block_hash,
                 "height": height,
-                "timestamp": datetime.utcfromtimestamp(block["time"]).isoformat(),
-                "mining_pool": mining_pool,
-                "analysis": analysis
+                "timestamp": datetime.utcfromtimestamp(doc["timestamp"]).isoformat(),
+                "mining_pool": doc.get("mining_pool"),
+                "analysis": doc.get("analysis")
             }
         }
         publish_to_rabbitmq(rabbitmq_doc)
     except Exception as e:
         logger.error("Error processing block %s: %s", block_hash, e)
 
+def reprocess_block_full(block_hash: str, publish_update: bool = False):
+    try:
+        doc = build_block_doc(block_hash)
+        height = doc["height"]
+        blocks_coll.replace_one({"block_hash": block_hash}, doc, upsert=True)
+        logger.info("Reindexed block %d (%s) with full overwrite", height, block_hash)
+        if publish_update:
+            rabbitmq_doc = {
+                "type": "block",
+                "id": str(uuid.uuid4()),
+                "timestamp": datetime.utcnow().isoformat(),
+                "data": {
+                    "hash": block_hash,
+                    "height": height,
+                    "timestamp": datetime.utcfromtimestamp(doc["timestamp"]).isoformat(),
+                    "mining_pool": doc.get("mining_pool"),
+                    "analysis": doc.get("analysis")
+                }
+            }
+            publish_to_rabbitmq(rabbitmq_doc)
+    except Exception as e:
+        logger.error("Error fully reprocessing block %s: %s", block_hash, e)
+
+def start_full_reindex_background():
+    try:
+        if not mongodb_enabled or db is None:
+            logger.warning("Historical data disabled or DB unavailable; skipping full reindex")
+            return
+        logger.info("Starting full background reindex of all blocks in descending height order")
+        count = 0
+        cursor = db.blocks.find({}, {"block_hash": 1, "height": 1, "_id": 0}).sort("height", -1)
+        for doc in cursor:
+            bhash = doc.get("block_hash")
+            if not bhash:
+                continue
+            reprocess_block_full(bhash)
+            count += 1
+            if count % 100 == 0:
+                logger.info(f"Reindexed {count} blocks so far (descending order)")
+        logger.info(f"Completed full background reindex of {count} blocks")
+    except Exception as e:
+        logger.error(f"Error starting full background reindex: {e}")
+
 from analytics.prev_hash_divergence import analyze_prev_hash_divergence
 from analytics.invalid_coinbase_no_merkle import analyze_invalid_coinbase_without_merkle
-from analytics.pool_identification import analyze_pool_identification, identify_pool_from_data, PoolsManager
-
-# Create the pools manager (must exist before startup_event and analysis usage)
-pools_manager = PoolsManager(db, POOL_LIST_URL, LOCAL_POOL_FILE, publish_to_rabbitmq)
+from analytics.pool_identification import analyze_pool_identification, identify_pool_from_data, load_pools
 
 def run_block_analyses(height: int, coinbase_script_hex: str | None = None, coinbase_addresses: List[str] | None = None) -> Dict[str, Any]:
     flags: List[Dict[str, Any]] = []
@@ -728,7 +352,7 @@ def run_block_analyses(height: int, coinbase_script_hex: str | None = None, coin
     if coinbase_script_hex is not None and coinbase_addresses is not None:
         try:
             analysis["pool_identification"] = analyze_pool_identification(
-                coinbase_script_hex, coinbase_addresses, pools_manager, logger
+                coinbase_script_hex, coinbase_addresses, pools_cache, logger
             )
         except Exception as e:
             logger.error(f"Error during pool identification analysis for height {height}: {e}")
@@ -737,6 +361,7 @@ def run_block_analyses(height: int, coinbase_script_hex: str | None = None, coin
         analysis["flags"] = flags
     return analysis
 
+
 def sync_blocks():
     """
     Sync blocks from the Bitcoin node
@@ -744,11 +369,11 @@ def sync_blocks():
     try:
         # Get the current best block
         best_block_hash = retry_rpc(
-            get_best_block_hash, 
+            get_best_block_hash,
             "getbestblockhash"
         )
         best_block = retry_rpc(
-            get_block, 
+            get_block,
             f"getblock({best_block_hash})",
             best_block_hash
         )
@@ -776,9 +401,6 @@ def sync_blocks():
         if lowest_processed_height is not None and lowest_processed_height > MIN_BLOCK_HEIGHT:
             logger.info(f"Checking for missing blocks between MIN_BLOCK_HEIGHT ({MIN_BLOCK_HEIGHT}) and lowest processed block ({lowest_processed_height})")
             
-            # Find missing blocks in the range using MongoDB's aggregation
-            missing_blocks = []
-            
             # First, get all heights in the range that are already processed
             processed_heights_in_range = set(
                 doc["height"] for doc in db.blocks.find(
@@ -795,7 +417,7 @@ def sync_blocks():
             
             if missing_blocks:
                 logger.info(f"Found {len(missing_blocks)} missing blocks between MIN_BLOCK_HEIGHT and lowest processed block")
-                if len(missing_blocks) < 20:  # Only log all missing blocks if there aren't too many
+                if len(missing_blocks) < 20:
                     logger.info(f"Missing blocks: {missing_blocks}")
                 else:
                     logger.info(f"First 10 missing blocks: {missing_blocks[:10]}")
@@ -813,17 +435,17 @@ def sync_blocks():
                     for height in missing_blocks[1:]:
                         if height != prev_height + 1:
                             # End of a consecutive range
-                            ranges.append((prev_height, range_start))  # Note: reversed for sync_range (high to low)
+                            ranges.append((prev_height, range_start))  # reversed for sync_range (high to low)
                             range_start = height
                         prev_height = height
                     
                     # Add the last range
-                    ranges.append((prev_height, range_start))  # Note: reversed for sync_range (high to low)
+                    ranges.append((prev_height, range_start))  # reversed for sync_range (high to low)
                     
                     logger.info(f"Identified {len(ranges)} consecutive ranges of missing blocks")
                     
                     # Process each range using sync_range
-                    for range_end, range_start in ranges:  # Reversed order for sync_range
+                    for range_end, range_start in ranges:  # reversed order for sync_range
                         logger.info(f"Processing range from height {range_end} down to {range_start}")
                         sync_range(range_end, range_start)
                 else:
@@ -837,7 +459,7 @@ def sync_blocks():
                             try:
                                 # Get the block hash for this height
                                 block_hash = retry_rpc(
-                                    get_block_hash, 
+                                    get_block_hash,
                                     f"getblockhash({height})",
                                     height
                                 )
@@ -868,12 +490,13 @@ def sync_blocks():
     except Exception as e:
         logger.error(f"Error during block sync: {e}")
 
+
 def sync_range(start_height, end_height):
     """
     Sync blocks in the specified range (from start_height down to end_height)
     """
     # Process blocks in small batches to avoid overwhelming the node
-    batch_size = 5  # Reduced batch size for better stability
+    batch_size = 5
     
     for batch_start in range(start_height, end_height - 1, -batch_size):
         batch_end = max(batch_start - batch_size + 1, end_height)
@@ -884,7 +507,7 @@ def sync_range(start_height, end_height):
             try:
                 # Get the block hash for this height
                 block_hash = retry_rpc(
-                    get_block_hash, 
+                    get_block_hash,
                     f"getblockhash({height})",
                     height
                 )
@@ -901,7 +524,7 @@ def sync_range(start_height, end_height):
                 process_block(block_hash, is_new_block=False)
                 
                 # Add a small delay between blocks to avoid overwhelming the node
-                time.sleep(0.5)  # Increased delay between blocks
+                time.sleep(0.5)
                 
             except Exception as e:
                 logger.error(f"Error processing block at height {height}: {e}")
@@ -910,10 +533,39 @@ def sync_range(start_height, end_height):
         
         # Add a delay between batches to let the node breathe
         logger.info(f"Completed batch from {batch_start} to {batch_end}, sleeping before next batch")
-        time.sleep(5)  # Increased delay between batches
+        time.sleep(5)
         
         # Recreate the connection pool between batches to ensure fresh connections
         rpc_pool.recreate_pool()
+
+def load_pools_once():
+    global pools_cache, pools_hash
+    try:
+        pools, new_hash, _ = load_pools(db, POOL_LIST_URL, LOCAL_POOL_FILE, pools_hash)
+        pools_cache = pools or {}
+        pools_hash = new_hash
+        logger.info(f"Initialized pools cache with {len(pools_cache)} entries")
+    except Exception as e:
+        logger.error(f"Failed to initialize pools cache: {e}")
+
+
+def pool_updater_task():
+    """Background task to periodically check for pool definition updates"""
+    while True:
+        try:
+            logger.info("Checking for updates to mining pool definitions")
+            global pools_cache, pools_hash
+            pools, new_hash, changed = load_pools(db, POOL_LIST_URL, LOCAL_POOL_FILE, pools_hash)
+            if pools:
+                pools_cache = pools
+                if new_hash is not None and new_hash != pools_hash:
+                    pools_hash = new_hash
+                    logger.info("Pool definitions changed; triggering background full reindex")
+                    threading.Thread(target=start_full_reindex_background, daemon=True).start()
+            time.sleep(POOL_UPDATE_INTERVAL)
+        except Exception as e:
+            logger.error(f"Error in pool updater task: {e}")
+            time.sleep(3600)  # Retry after an hour
 
 def zmq_listener():
     context = zmq.Context()
@@ -936,8 +588,8 @@ def zmq_listener():
                     # Get block hash in correct little-endian format
                     bhash = block.GetHash()[::-1].hex()
                     logger.info("ZMQ received new block: %s", bhash)
-                    # Submit new block to the dedicated new block processor
-                    new_block_processor.submit(process_block, bhash, True)
+                    # Submit new block to the block processor
+                    block_processor.submit(process_block, bhash, True)
                 except Exception as inner_e:
                     logger.error("Error processing ZMQ message: %s", inner_e)
                     time.sleep(1)
@@ -945,117 +597,53 @@ def zmq_listener():
             logger.error("ZMQ connection error: %s", outer_e)
             time.sleep(5)
 
-def pool_updater_task():
-    """Background task to periodically check for pool definition updates"""
-    while True:
-        try:
-            logger.info("Checking for updates to mining pool definitions")
-            pools_manager.load_pools()
-            
-            # Sleep for configured interval
-            time.sleep(POOL_UPDATE_INTERVAL)
-        except Exception as e:
-            logger.error(f"Error in pool updater task: {e}")
-            time.sleep(3600)  # Retry after an hour
-
 # --- RabbitMQ Heartbeat Task ---
 def rabbitmq_heartbeat_task():
-    """
-    Periodically check and maintain the RabbitMQ connection
-    """
-    HEARTBEAT_INTERVAL = 15  # seconds
-    
-    logger.info("Starting RabbitMQ heartbeat task")
-    
-    while True:
-        try:
-            # Check if connection is active and process any pending events
-            if rabbitmq_manager.connection and not rabbitmq_manager.connection.is_closed:
-                try:
-                    # Process any pending events to keep the connection alive
-                    rabbitmq_manager.connection.process_data_events(time_limit=0)
-                    logger.debug("RabbitMQ heartbeat check successful")
-                except Exception as e:
-                    logger.warning(f"RabbitMQ heartbeat check failed: {e}")
-                    # Force reconnection
-                    rabbitmq_manager.connect()
-            else:
-                # Connection is closed or doesn't exist, try to reconnect
-                logger.warning("RabbitMQ connection is closed during heartbeat check, reconnecting")
-                rabbitmq_manager.connect()
-                
-        except Exception as e:
-            logger.error(f"Error in RabbitMQ heartbeat task: {e}")
-            
-        time.sleep(HEARTBEAT_INTERVAL)
+    # Backward compatibility shim; delegate to integration module
+    from integrations.rabbitmq import heartbeat_task
+    return heartbeat_task()
 
-@app.on_event("startup")
-def startup_event():
-    """Initialize background tasks on startup"""
-    logger.info("Starting up Bitcoin Mining Pool Identification Service")
-    
-    # Load pool definitions
-    pools_manager.load_pools()
-    
-    # Start background threads
-    sync_thread = threading.Thread(target=sync_blocks, daemon=True)
-    sync_thread.start()
-    
-    zmq_thread = threading.Thread(target=zmq_listener, daemon=True)
-    zmq_thread.start()
-    
-    pool_updater_thread = threading.Thread(target=pool_updater_task, daemon=True)
-    pool_updater_thread.start()
-    
-    # Start RabbitMQ heartbeat thread
-    rabbitmq_heartbeat_thread = threading.Thread(target=rabbitmq_heartbeat_task, daemon=True)
-    rabbitmq_heartbeat_thread.start()
-    
-    # Note: We've removed the automatic reindexing that was here
-    # If you need to reindex blocks, use the --reindex-blocks command line argument
-    
-    logger.info("Application startup complete")
+def start_background_workers():
+    """Initialize background tasks when running as a worker-only service."""
+    logger.info("Starting background workers for Bitcoin Mining Pool Identification Service")
+    load_pools_once()
+    threading.Thread(target=sync_blocks, daemon=True).start()
+    threading.Thread(target=zmq_listener, daemon=True).start()
+    threading.Thread(target=pool_updater_task, daemon=True).start()
+    start_heartbeat_thread()
+    if START_FULL_REINDEX_ON_BOOT:
+        logger.info("Starting full background reindex as requested by CLI flag")
+        threading.Thread(target=start_full_reindex_background, daemon=True).start()
+    logger.info("Background workers started")
 
-@app.on_event("shutdown")
-async def shutdown_event():
-    logger.info("Application shutdown: cleaning up resources")
-    
-    # Close RabbitMQ connection
+def shutdown_cleanup():
+    logger.info("Shutdown: cleaning up resources")
     try:
         if rabbitmq_manager.connection and not rabbitmq_manager.connection.is_closed:
             logger.info("Closing RabbitMQ connection")
             rabbitmq_manager.connection.close()
     except Exception as e:
         logger.error(f"Error closing RabbitMQ connection: {e}")
-    
-    old_block_processor.shutdown(wait=True)
-    new_block_processor.shutdown(wait=True)
+    block_processor.shutdown(wait=True)
     logger.info("Thread pools shut down")
 
 if __name__ == "__main__":
     import argparse
-    import uvicorn.config
-    
-    parser = argparse.ArgumentParser(description='Block indexer service')
-    parser.add_argument('--update-pools', action='store_true', 
-                        help='Update pool definitions and reindex blocks')
+    parser = argparse.ArgumentParser(description='Block indexer worker service')
     parser.add_argument('--reindex-blocks', action='store_true',
                         help='Force reindexing of all blocks')
     args = parser.parse_args()
     
     # Handle command-line options
-    if args.update_pools:
-        logger.info("Manual update of pool definitions requested via command line")
-        pools_manager.load_pools()
-        
     if args.reindex_blocks:
-        logger.info("Manual reindexing of blocks requested via command line")
-        pools_manager.reindex_blocks()
+        logger.info("Manual full reindex requested; will run in background after startup")
+        # Set flag to trigger background full reindex on app startup
+        START_FULL_REINDEX_ON_BOOT = True
     
-    # Get port from environment variable
-    port = int(os.getenv("PORT", "8001"))
-    host = "0.0.0.0"  # Explicitly bind to all interfaces
-    logger.info(f"Starting FastAPI server on {host}:{port}")
-    config = uvicorn.config.Config(app, host=host, port=port, loop="asyncio")
-    server = uvicorn.Server(config)
-    server.run() 
+    # Start background workers instead of HTTP server
+    start_background_workers()
+    try:
+        while True:
+            time.sleep(60)
+    except KeyboardInterrupt:
+        shutdown_cleanup()
