+++ backend/requirements.txt
@@ -1,5 +1,3 @@
-fastapi==0.95.2
-uvicorn==0.22.0
 pyzmq==25.1.1
 pymongo==4.7.2
 python-bitcoinrpc==1.0
